shared memory for Coulomb density
Defined groups. Now Orbitals are in an own group, processes that share the same memory are also identified. The rank in the orbital group are consecutive if they are in the same shared memory group.
Replaced MPI_rank by MPI_Orb_rank.

Densities and SerialTrees hav now an additional flag that tell if the coefficients are stored in shared memory.
The shared density is now implemented only for total densities. (not "Spin, Alpha or Beta").

Also redefined ScaledPrecision and phi_jij/phi_iij (see issue 35)