Do not copy internal header files to build directory
Include paths for targets built within MRCPP have both `<build-dir>/include`
and `<src-dir>/src`.
This is needed to ensure that the API headers are used uniformly _inside_ and
_outside_ the project, that is: `#include "MRCPP/MWFunctions"` works both when
working in MRCPP and when working in MRChem or VAMPyR.
To use the internal headers, one needs to either mirror the structure of the
installation within the build directory or to add `<src-dir>/src` to the include
paths. I was doing the mirroring _and_ the extension of the include paths, but
this confused Clang (surprisingly not GCC!) that didn't know which header to
actually include.
Solution: add `<src-dir>/src` to the include paths and only mirror the API
headers to `<build-dir>/include`. The API headers are now copied to the build
directory using `configure_file`, which ensures that they are copied fresh
anytime they are changed.