Add negative scale operator and reach/cut_off operators.
The negative scale operator generates parent nodes, it later
uses in the operator application to increase the reach of the operator.

For zeroth scale operators we can use a reach and a cut off parameter
to determine the truncation of the operators. In addition to this we
can split the applies into a near field (inside unit-cell) and a far field
(outside unit-cell).